kernel: fitimage: deploy real fit-image
Currently, do_deploy installs a standard kernel image as an ITB. This
image is linux.bin and it is prepared by uboot_prep_kimage.

The real kernel fit-image is build under arch/${ARCH}/boot/fitImage.

This patch deploys the real fit-image instead of the one prepared by
u-boot and which is included in the ITB.

Note: this patch only concerns fit-images without initramfs.

Signed-off-by: Gaël PORTAY <gael.portay@savoirfairelinux.com>